set up project header with props and color transitions
